aws-integration/requirements.txt
boto3==1.28.25
python-dotenv==1.0.0
pandas==2.0.3

